requirements.txt
pandas==1.3.0
argparse==1.4.0
azure-common==1.1.27
azure-core==1.18.0
azure-identity==1.6.1
azure-keyvault==4.1.0
azure-keyvault-administration==4.0.0
azure-keyvault-certificates==4.3.0
azure-keyvault-keys==4.4.0
azure-keyvault-secrets==4.3.0
azure-storage-blob==12.8.1
azure-storage-blob==12.8.1
openpyxl==3.0.9
numpy==1.23.0
pymongo==4.1.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
azure-mgmt-cosmosdb==6.4.0
SQLAlchemy==1.4.23
sqlitedict==1.7.0
pytictoc==1.5.2
pathvalidate==2.5.1